refactor(MenuService): 금액 검증 로직을 상품이 하도록 변경
