+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,6 +1,5 @@
 package kitchenpos.menu.application;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.menu.dao.MenuProductRepository;
@@ -11,34 +10,33 @@ import kitchenpos.menu.domain.MenuProducts;
 import kitchenpos.menu.dto.MenuProductRequest;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
-import kitchenpos.product.dao.ProductRepository;
-import kitchenpos.product.domain.Product;
+import kitchenpos.product.application.ProductService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class MenuService {
     private final MenuRepository menuRepository;
-    private final ProductRepository productRepository;
     private final MenuGroupService menuGroupService;
     private final MenuProductRepository menuProductRepository;
+    private final ProductService productService;
 
     public MenuService(
             final MenuRepository menuRepository,
-            final ProductRepository productRepository,
             final MenuGroupService menuGroupService,
-            final MenuProductRepository menuProductRepository
+            final MenuProductRepository menuProductRepository,
+            final ProductService productService
     ) {
         this.menuRepository = menuRepository;
-        this.productRepository = productRepository;
         this.menuGroupService = menuGroupService;
         this.menuProductRepository = menuProductRepository;
+        this.productService = productService;
     }
 
     @Transactional
     public MenuResponse create(final MenuRequest menuRequest) {
         menuGroupService.existsById(menuRequest.getMenuGroupId());
-        validPriceCheck(menuRequest.getPrice(), menuRequest.getMenuProducts());
+        productService.validPriceCheck(menuRequest);
         Menu savedMenu = menuRepository.save(menuRequest.toMenu());
 
         MenuProducts menuProducts = createMenuProduct(menuRequest.getMenuProducts());
@@ -55,20 +53,6 @@ public class MenuService {
         return menus.stream()
                 .map(MenuResponse::of)
                 .collect(Collectors.toList());
-    }
-
-    private void validPriceCheck(BigDecimal price, List<MenuProductRequest> menuProductRequestList) {
-        BigDecimal sum = BigDecimal.ZERO;
-
-        for (final MenuProductRequest menuProductRequest : menuProductRequestList) {
-            final Product product = productRepository.findById(menuProductRequest.getProductId())
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 상품입니다."));
-            sum = sum.add(product.multiplyPrice(menuProductRequest.getQuantity()));
-        }
-
-        if (price.compareTo(sum) > 0) {
-            throw new IllegalArgumentException("메뉴의 금액은 상품의 합 보다 작아야합니다.");
-        }
     }
 
     private MenuProducts createMenuProduct(List<MenuProductRequest> menuProductRequestList) {
