+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,6 +1,9 @@
 package kitchenpos.product.application;
 
+import java.math.BigDecimal;
 import java.util.stream.Collectors;
+import kitchenpos.menu.dto.MenuProductRequest;
+import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.product.dao.ProductRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.dto.ProductRequest;
@@ -29,5 +32,20 @@ public class ProductService {
         return productRepository.findAll().stream()
                 .map(ProductResponse::of)
                 .collect(Collectors.toList());
+    }
+
+    @Transactional(readOnly = true)
+    public void validPriceCheck(final MenuRequest menuRequest) {
+        BigDecimal sum = BigDecimal.ZERO;
+
+        for (final MenuProductRequest menuProductRequest : menuRequest.getMenuProducts()) {
+            final Product product = productRepository.findById(menuProductRequest.getProductId())
+                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 상품입니다."));
+            sum = sum.add(product.multiplyPrice(menuProductRequest.getQuantity()));
+        }
+
+        if (sum.compareTo(menuRequest.getPrice()) < 0) {
+            throw new IllegalArgumentException("메뉴의 금액은 상품의 합 보다 작아야합니다.");
+        }
     }
 }
