+++ src/test/java/kitchenpos/product/ProductServiceTest.java
@@ -8,7 +8,11 @@ import static org.mockito.BDDMockito.given;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import kitchenpos.menu.dto.MenuProductRequest;
+import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.dao.ProductRepository;
 import kitchenpos.product.domain.Product;
@@ -33,6 +37,7 @@ class ProductServiceTest {
 
     Product 후라이드;
     Product 양념치킨;
+    MenuRequest 상품;
 
     @BeforeEach
     void setUp() {
@@ -41,8 +46,10 @@ class ProductServiceTest {
 
     void createProduct() {
         후라이드 = new Product("후라이드", BigDecimal.valueOf(15000));
-
         양념치킨 = new Product("양념치킨", BigDecimal.valueOf(15000));
+
+        상품 = new MenuRequest("후라이드치킨", BigDecimal.valueOf(15000), 1L,
+                Collections.singletonList(new MenuProductRequest(1L, 1L)));
     }
 
     @Test
@@ -85,5 +92,45 @@ class ProductServiceTest {
 
         // then
         assertThat(actual).hasSize(2);
+    }
+
+
+    @Test
+    @DisplayName("메뉴 저장시 메뉴는 존재하는 상품 정보를 가져야 한다")
+    void create_nonProductInfoError() {
+        // given
+        given(productRepository.findById(any())).willReturn(Optional.empty());
+
+        // when & then
+        assertThatIllegalArgumentException().isThrownBy(
+                () -> productService.validPriceCheck(상품)
+        ).withMessageContaining("존재하지 않는 상품입니다.");
+    }
+
+    @Test
+    @DisplayName("메뉴 저장시 메뉴상품에 속한 상품들의 금액 합보다 메뉴 가격이 작아야 한다")
+    void create_totalPriceError() {
+        // given
+        상품 = new MenuRequest("후라이드치킨", BigDecimal.valueOf(20000), 1L,
+                Collections.singletonList(new MenuProductRequest(1L, 1L)));
+        given(productRepository.findById(any())).willReturn(Optional.ofNullable(후라이드));
+
+        // when & then
+        assertThatIllegalArgumentException().isThrownBy(
+                () -> productService.validPriceCheck(상품)
+        ).withMessageContaining("메뉴의 금액은 상품의 합 보다 작아야합니다.");
+    }
+
+    @Test
+    @DisplayName("메뉴 저장시 메뉴의 금액은 0원 이상이다")
+    void create_priceException() {
+        // given
+        상품 = new MenuRequest("후라이드치킨", BigDecimal.valueOf(-1), 1L,
+                Collections.singletonList(new MenuProductRequest(1L, 1L)));
+
+        // when & then
+        assertThatIllegalArgumentException().isThrownBy(
+                () -> productService.validPriceCheck(상품)
+        );
     }
 }
